thrift refactor

+++ app/web/hapi/serve.go
@@ -50,7 +50,7 @@ func beforeHanding(h echo.HandlerFunc) echo.HandlerFunc {
 }
 
 //检查会员编号
-func getMemberId(c *echox.Context) int64 {
+func getMemberId(c *echox.Context) int {
 	return shared.GetMemberId(c)
 }
